requirement.txt
fastapi>=0.110.0
python-dotenv
uvicorn>=0.29.0
groq>=0.5.0
chromadb>=0.5.0
pydantic>=2.7.0
# Tree-sitter core and Python grammar binding
tree-sitter>=0.24.0
tree-sitter-python>=0.23.0
python-multipart